% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesMvLMconjugate.R
\name{pred_bayesMvLMconjugate}
\alias{pred_bayesMvLMconjugate}
\title{Predictive sampler for Conjugate Bayesian Multivariate Linear Models}
\usage{
pred_bayesMvLMconjugate(X_new, B_samples, Sigma_samples)
}
\arguments{
\item{X_new}{\link{matrix} \eqn{n_new \times p} of predictors for new data points}

\item{B_samples}{\link{array} of posterior sample for \eqn{\beta}}

\item{Sigma_samples}{\link{array} of posterior samples for \eqn{\Sigma}}
}
\value{
Y_pred \link{matrix} of posterior mean for response matrix Y predictions

Y_pred_samples \link{array} of posterior predictive sample for response matrix Y
}
\description{
Predictive sampler for Conjugate Bayesian Multivariate Linear Models
}
