% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spBPS-package.R
\docType{package}
\name{spBPS-package}
\alias{spBPS-package}
\title{spBPS: Bayesian Predictive Stacking for Scalable Geospatial Transfer Learning}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functions for Bayesian Predictive Stacking within the Bayesian transfer learning framework for geospatial artificial systems, as introduced in "Bayesian Transfer Learning for Artificially Intelligent Geospatial Systems: A Predictive Stacking Approach" (Presicce and Banerjee, 2025) \doi{10.48550/arXiv.2410.09504}. This methodology enables efficient Bayesian geostatistical modeling, utilizing predictive stacking to improve inference across spatial datasets. The core functions leverage 'C++' for high-performance computation, making the framework well-suited for large-scale spatial data analysis in parallel and distributed computing environments. Designed for scalability, it allows seamless application in computationally demanding scenarios.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lucapresicce.github.io/spBPS/}
}

}
\author{
\strong{Maintainer}: Luca Presicce \email{l.presicce@campus.unimib.it} (\href{https://orcid.org/0009-0005-7062-3523}{ORCID})

Authors:
\itemize{
  \item Sudipto Banerjee
}

}
\keyword{internal}
